refactor(mastracode): split run-checks + repo-cleanup, document duplicate-named module pairs
Phase 2 — split tool files that had outgrown their single-responsibility scope:

run-checks.ts (484 → 169 lines) split into four focused modules:
  • check-runner.ts      — runner detection, command building, subprocess execution
  • check-parsers.ts     — parser-registry wrapper + SHA256 fingerprinting
  • check-convergence.ts — converging/stalled/resolved iteration state
  • run-checks.ts        — Mastra tool definition only

repo-cleanup.ts (315 → 197 lines) split into three focused modules:
  • cleanup-report.ts    — shadow-scan report extraction + validation
  • cleanup-fixes.ts     — apply delete/move/gitignore remediations
  • repo-cleanup.ts      — Mastra tool definition (orchestrator)

Phase 3 — add 2-line header comments to the four pairs of duplicate-named
modules so it's instantly obvious which file is the data layer and which is
the wrapper:

  • src/confidence-journal.ts   ↔ src/tools/confidence-journal.ts
  • src/session-ledger.ts        ↔ src/tools/session-ledger.ts
  • src/verification-result.ts   ↔ src/tools/verification-result.ts
  • src/shadow-scanner.ts        ↔ src/subagents/shadow-scanner.ts

The data-layer files already had good headers; the wrapper files now point
back to them. No behavior changes — pure mechanical extraction. tsc clean,
eslint clean, build succeeds, harness boots.
